types again

+++ tables.sql
@@ -1,6 +1,6 @@
 CREATE TABLE dht (
 	epoch int(11) not null,
-	humidity float(2,2) not null,
-	temperature float(2,2) not null,
-	heatindex float(2,2) not null
+	humidity decimal(4,2) not null,
+	temperature decimal(4,2) not null,
+	heatindex decimal(4,2) not null
 );
